+++ code/FinanceManager.Components/Models/LoginModel.cs
@@ -5,7 +5,7 @@ namespace FinanceManager.Components.Models;
 public class LoginModel
 {
     [Required(AllowEmptyStrings = false)]
-    [StringLength(20, MinimumLength = 3, ErrorMessage = "Login must be between 3 and 20 characters")]
+    [StringLength(254, MinimumLength = 5, ErrorMessage = "Email must be between 5 and 254 characters")]
     public string? Login { get; set; }
 
     [Required(AllowEmptyStrings = false)]