+++ code/FinanceManager.Domain/Commands/User/AddUser.cs
@@ -2,4 +2,4 @@
 
 namespace FinanceManager.Application.Commands.User;
 
-public record AddUser(string UserName, string Password, PricingLevel PricingLevel);
+public record AddUser(string UserName, string Password, PricingLevel PricingLevel, string? FirstName = null, string? LastName = null);